added new items

+++ Entities/Enemies/ParentEnemy.cs
@@ -30,7 +30,7 @@ namespace DungeonExplorer
                 int attack = (int)attackDecimal;
                 int defence = player.CurrentDefence;
 
-                double damageReduction = defence / (defence + 50);
+                double damageReduction = (double)defence / ((double)defence + 50);
                 double damageDouble = attack * (1 - damageReduction);
                 int damage = (int)damageDouble;
